move isAllowed function to gitea/utils
